:sparkles: feat: 태그 엔티티 수정 및 조회 로직 개선
- `Tag` 엔티티 수정
- 태그 조회 로직 변경

Related to: #64

+++ src/main/java/com/example/festimo/domain/post/dto/TagResponse.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class TagResponse {
     private String name;
-    private Long count;
+    private int count;
 }
